Add models and basic functionality
This commit adds `PriceLevel` which is the most granular data structure
in the order book. This `PriceLevel` contains orders on one `Price` on
one `Side`.
`PriceLevel`'s are contained in a `BookSide` which represents a single
side of an order book.
Finally we have the `OrderBook` which holds exactly 2 `BookSide`'s.